Update carp.c
Fire the _upscript_ in case of a preferred master is reasserting control, too. The master will send gratuitous arp for the primary VIP. But there may be more then one provided by the up script. Also, it might be need to perform some actions if another instance have be become a master temporary and fired the upscript there.